Implement combination sum algorithm
Problem:
Given an array of distinct integers nums and a target integer target, return a list of all unique combinations of nums where the chosen numbers sum to target. You may return the combinations in any order.
The same number may be chosen from nums an unlimited number of times. Two combinations are unique if the frequency of at least one of the chosen numbers is different.
The test cases are generated such that the number of unique combinations that sum up to the target is less than 150 combinations for the given input.

Constraints
1≤nums.length≤30
1≤nums[i]≤40
All elements of nums are distinct.
1≤target≤40